requirements.txt
selenium==4.11.0
black==23.7.0
Pillow 
reportlab
pytesseract 
pdf2image 
PyPDF2
